rename host package to processor
