+++ beanstalk/worker.go
@@ -1,12 +1,10 @@
 package beanstalk
 
 import (
-	"encoding/json"
 	"log"
 	"time"
 
 	"github.com/iwanbk/gobeanstalk"
-	"github.com/robin-moser/bugspider/host"
 )
 
 func (bs *Handler) Watch(tubes ...string) error {
@@ -22,31 +20,6 @@ func (bs *Handler) Watch(tubes ...string) error {
 	}
 	log.Println("watching", watching, "tubes")
 	return nil
-}
-
-func (bs *Handler) ProcessJob(processor host.Processor) {
-	job, err := bs.serverConnection.Reserve()
-	if err != nil {
-		log.Println(err)
-		return
-	}
-
-	currentHost := host.Host{}
-	err = json.Unmarshal(job.Body, &currentHost)
-	if err != nil {
-		bs.handleError(job, err)
-		return
-	}
-
-	inserted, err := processor.Process(&currentHost)
-	if err != nil {
-		bs.handleError(job, err)
-		return
-	}
-	if inserted {
-		log.Printf("processed Job ID %v: saved %v\n", job.ID, currentHost.Hostname)
-	}
-	bs.serverConnection.Delete(job.ID)
 }
 
 // handleError gets called, when a job cant finish, so it can be released
