+++ scraper/scraper.go
@@ -3,39 +3,41 @@ package scraper
 import (
 	"errors"
 
-	"github.com/robin-moser/bugspider/host"
+	"github.com/robin-moser/bugspider/processor"
 	"github.com/robin-moser/bugspider/scraper/immuniweb"
 	"github.com/robin-moser/bugspider/scraper/ssllabs"
 )
 
 // Scrape domains from various Sources.
 // A predefined source needs to be given.
-func Scrape(source string) (*host.Collection, error) {
+func Scrape(source string) (*processor.Collection, error) {
 
 	switch source {
 	case "immuniweb":
-		hostarray, err := immuniweb.Scrape()
-		if err != nil {
-			return nil, err
-		} else if len(hostarray.Hosts) == 0 {
-			return nil, errors.New("Error: no domains found")
-		}
-
-		return hostarray, nil
+		return populateHosts(immuniweb.Scrape())
 
 	case "ssllabs":
-		hostarray, err := ssllabs.Scrape()
-		if err != nil {
-			return nil, err
-		} else if len(hostarray.Hosts) == 0 {
-			return nil, errors.New("Error: no domains found")
-		}
-
-		return hostarray, nil
+		return populateHosts(ssllabs.Scrape())
 
 	default:
 		err := errors.New("Undefined Scrape Source")
 		return nil, err
 	}
 
+}
+
+func populateHosts(hostarray *processor.Collection, err error) (*processor.Collection, error) {
+	if err != nil {
+		return nil, err
+	}
+
+	if len(hostarray.Hosts) == 0 {
+		return nil, errors.New("Error: no domains found")
+	}
+
+	for i := 0; i < len(hostarray.Hosts); i++ {
+		hostarray.Hosts[i].JobType = "deduplication"
+	}
+
+	return hostarray, nil
 }
